anarchy-copilot/tests/requirements-test.txt
# Test Dependencies
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-xdist>=2.5.0
pytest-timeout>=2.1.0
pytest-mock>=3.6.0
beautifulsoup4>=4.10.0
jinja2>=3.0.0
pyyaml>=6.0.0
httpx>=0.22.0
aiohttp>=3.8.0
types-aiohttp>=3.8.0
mypy>=0.950
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
coverage>=6.3.0

# Runtime dependencies needed for testing
fastapi>=0.68.0
sqlalchemy>=1.4.0
pydantic>=1.9.0
rich>=12.0.0
typer>=0.4.0

# Mock dependencies
aioresponses>=0.7.0
responses>=0.17.0
mock>=4.0.0

# Type stubs
types-PyYAML
types-beautifulsoup4
types-requests
types-setuptools
types-mock

# Development tools
pre-commit>=2.17.0
bandit>=1.7.0
safety>=1.10.0
pip-audit>=2.0.0

# Documentation generation
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.12.0

# Additional utilities
freezegun>=1.2.0  # For time-based testing
tenacity>=8.0.0   # For retry testing
hypothesis>=6.0.0 # For property-based testing
